Splitting code into multiple files, _() cleanup
