+++ AUTHORS
@@ -1,0 +1,13 @@
+purple-mattermost libpurple plugin
+
+Author:
+
+Eion Robb <eion@robbmob.com>
+
+Contributors:
+
+Edvard Rejthar <github@edvard.cz>
+it33 <it33@users.noreply.github.com>
+Jarek Polok <jaroslaw.polok@gmail.com>
+Patrick Laxton <patrick.laxton@protonmail.com>
+Peter Chubb <Peter.Chubb@data61.csiro.au>
